requirements.txt
Flask
mysql-connector
flask-pyjwt